+++ packages/hardhat-helpers/src/index.js
@@ -2,10 +2,12 @@ const { etherscan } = require('./etherscan')
 const { networks } = require('./networks')
 const tasks = require('./tasks')
 const gov = require('./gov')
+const balance = require('./balance')
 
 module.exports = {
   networks,
   etherscan,
   tasks,
   ...gov,
+  ...balance,
 }
